Telefone Mapeamento

+++ Domain/Entities/Telefone.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities
     public class Telefone : IBase
     {
         public virtual int Id { get; set; }
-        public virtual int Numero { get; set; }
+        public virtual long Numero { get; set; }
         public virtual TipoTel TipoTel { get; set; }
     }
 }
